Adapt BMP085 driver to new i2c interface
Right now the barometric pressure data is not included in the sensor
message because after including the additional bytes the CRC check
failed in the basestation. In the current develop branch is ongoing work
to redefine the message format, after finalizing I'm happy to add a
patch to adapt it.

The barometric pressure data can be seen on the temperature sensor debug
port (just #define UART_DEBUG and recompile).

    Code size:
             text    data     bss     dec     hex
    BMP085: 12626     289     544   13459    3493
    v0.2.0: 10260     289     509   11058    2b32

Conflicts:
	firmware/shc_envsensor/shc_envsensor.c