bugfix: assemble state string correctly for dimmer device
The bugfix in commit id 30bd0677224 only fixes the state string creation for
the power switch, but breaks it for the dimmer.

The define statement for the dimmer is different from the power switch
because the state string also contains the brightness (e.g. "on B: 100")

Use define SHC_DIM_60_notify_on notify SHC_DIM_60:on.B:.100