Maybe multiple packets need to be sent
If the wakeup count for different measurements is reached at the same time, only one of them gets sent. F.ex. if you check the  temperature every 15 minutes, and the brightness every hour, at the fourth temperature measurement the brightness should be sent at the same time. The actual code checks if a temperature has to be sent, and if it does, the rest of the values is not checked at all.
Is this by purpose, to prevent the envsensor from sending too much packets?